Safari-609.1.11.1/LayoutTests/inspector/formatting/formatting-css-expected.txt

Test CSS formatting.


== Running test suite: CSSFormatter
-- Running test case: CSSFormatter
PASS: basic.css
PASS: comment.css
PASS: gradient.css
PASS: keyframes.css
PASS: media-query.css
PASS: selectors.css
PASS: wrapping.css
